requirements.txt
arrow==0.14.2
certifi==2019.6.16
chardet==3.0.4
idna==2.8
numpy==1.16.4
oauthlib==3.0.2
paho-mqtt==1.4.0
pandas==0.24.2
parse==1.12.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.25.3
